requirements.txt
coverage==4.0.3
Django==1.9
django-crispy-forms==1.6.0
django-nose==1.4.3
django-registration==2.1.2
easy-thumbnails==2.3
loginform==1.2.0
lxml==3.6.4
nose==1.3.7
Pillow==3.0.0
py==1.4.31
pytest==2.9.2
PyYAML==3.11
wheel==0.24.0
